add(Serializer.cs): вынес логику сериализации в отдельный класс
